Create Pixel-Tech.py
This is a rock-paper-scissor game. Random library has been used here and various functions have been created. 
The game starts with a greeting message and then asks the user to enter his/her name. After that, the user chooses his hand from the list provided. The computer chooses a random hand and then validates it. After validating, the result is printed (in the form of whether the user has won or lost).